part III

+++ src/com/participant/examen/Examen.java
@@ -2,23 +2,23 @@ package com.participant.examen;
 
 
 public class Examen {
-    private int notaExamen;
+    private Float notaExamen;
     private int salaExamen;
     private int idExamen;
 
-    public Examen(int idExamen, int salaExamen, int notaExamen) {
+    public Examen(int idExamen, int salaExamen, Float notaExamen) {
         this.notaExamen = notaExamen;
         this.salaExamen = salaExamen;
         this.idExamen = idExamen;
     }
 
-    public int getNotaExamen() { return notaExamen; }
+    public Float getNotaExamen() { return notaExamen; }
 
     public int getSalaExamen() { return salaExamen; }
 
     public int getIdExamen() { return idExamen; }
 
-    public void setNotaExamen(int notaExamen) { this.notaExamen = notaExamen; }
+    public void setNotaExamen(Float notaExamen) { this.notaExamen = notaExamen; }
 
     public void setSalaExamen(int salaExamen) { this.salaExamen = salaExamen; }
